color util added
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk@1654 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ jasperreports/src/net/sf/jasperreports/engine/export/oasis/Utility.java
@@ -60,7 +60,7 @@ public class Utility {
 		return inches;
 	}
 
-	public static double translatePixelsToInchesWithNoRoundOff(int pixels){
+	public static double translatePixelsToInchesWithNoRoundOff(double pixels){
 		double inches = 0.0;
 		inches = pixels/72.0;
 		return inches;
